updated installer to install fox to user space rather than system space

+++ install.sh
@@ -11,8 +11,8 @@ set -e
 
 # ----- variables using env
 
-FOX_PREFIX=/usr/local/bin
-FOX_VERSION=2.0.0
+FOX_PATH=$HOME/.fox
+FOX_VERSION=2.1.0
 
 # ----- functions
 
@@ -24,8 +24,9 @@ if [[ $(id -u) != 0 ]]; then
   if command -v sudo >/dev/null 2>&1; then
     SUDO="sudo"
   else
-    echo ":: sudo not installed - aborting ::"
-    exit 1
+    echo ":: sudo not installed or no admin privileges detected ::"
+    SUDO=""
+    # exit 1
   fi
 fi
 
@@ -46,40 +47,44 @@ else
   exit 1;
 fi
 
-# remove old legacy-samurai/fox files
+# remove old legacy-samurai/fox files if admin privileges
 
-echo ":: checking and removing legacy samurai/butlerfox files ::"
+if [[ "$SUDO" == "sudo" ]]; then
+  echo ":: checking and removing legacy samurai/butlerfox files ::"
 
-if [ -f /usr/local/bin/samurai ]; then
-  echo ":: found and removing legacy /usr/local/bin/samurai file ::"
-  $SUDO rm /usr/local/bin/samurai
+  if [ -f /usr/local/bin/samurai ]; then
+    echo ":: found and removing legacy /usr/local/bin/samurai file ::"
+    $SUDO rm /usr/local/bin/samurai
+  fi
+
+  if [ -f /usr/local/bin/samurai-mac ]; then
+    echo ":: found and removing legacy /usr/local/bin/samurai-mac file ::"
+    $SUDO rm /usr/local/bin/samurai-mac.py
+  fi
+
+  if [ -f /usr/local/bin/samurai-linux ]; then
+    echo ":: found and removing legacy /usr/local/bin/samurai-linux file ::"
+    $SUDO rm /usr/local/bin/samurai-linux.py
+  fi
 fi
 
-if [ -f /usr/local/bin/samurai-mac ]; then
-  echo ":: found and removing legacy /usr/local/bin/samurai-mac file ::"
-  $SUDO rm /usr/local/bin/samurai-mac.py
-fi
-
-if [ -f /usr/local/bin/samurai-linux ]; then
-  echo ":: found and removing legacy /usr/local/bin/samurai-linux file ::"
-  $SUDO rm /usr/local/bin/samurai-linux.py
-fi
 
 # install new project files
-echo ":: installing butler(fox) to $FOX_PREFIX ::"
-# example : curl -L -o master.zip http://github.com/zoul/Finch/zipball/master/
-# if install via github curl
-$SUDO curl -L https://raw.githubusercontent.com/mirageglobe/butlerfox/master/dist/fox-latest.sh -o $FOX_PREFIX/fox
+
+echo ":: installing butler(fox) to $FOX_PATH ::"
+mkdir -pv $FOX_PATH
+curl -L https://raw.githubusercontent.com/mirageglobe/butlerfox/master/dist/fox-latest.sh -o $FOX_PATH/fox
 
 echo ":: symlinking/setting butler(fox) ::"
-$SUDO chmod ug+x $FOX_PREFIX/fox
-command -v fox || { echo ":: failed to install fox at $FOX_PREFIX. please log issue at https://github.com/mirageglobe/butlerfox ::"; exit 1; }
+chmod u+x $FOX_PATH/fox
+command -v fox || { echo ":: failed to install fox at $FOX_PATH. please log issue at https://github.com/mirageglobe/butlerfox ::"; exit 1; }
 echo ":: complete. please restart shell for path to update ::"
 
 # summary
 
 echo ":: summary ::"
-echo "  installed butlerfox into $FOX_PREFIX/fox"
-# echo "  symlinked /usr/local/bin/samurai.sh to /opt/samurai.sh"
-echo "  to uninstall, delete binary $FOX_PREFIX/fox"
+echo "  installed butlerfox into $FOX_PATH/fox"
+echo "  add the following to your .bashrc or .zsh"
+echo "    export PATH=$FOX_PATH:$PATH"
+echo "  to uninstall, delete binary $FOX_PATH/fox"
